🎨: 장바구니 물건 계산 로직을 함수로 추출
- 장바구니의 제품 목록은 인자로 전달받고, 물건 개수와 할인 정보는 DOM 요소에 접근하여 총액 계산
